requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pipenv==2023.11.17
platformdirs==4.2.0
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
